myapi/requirements.txt
amqp==2.5.2
APScheduler==3.6.3
billiard==3.6.1.0
celery==4.3.0
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
importlib-metadata==0.23
kombu==4.6.6
more-itertools==7.2.0
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.3.11
six==1.13.0
sqlparse==0.3.0
tzlocal==2.0.0
vine==1.3.0
zipp==0.6.0
